+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added Vietnamese translation ([#124](https://codeberg.org/silverpill/mitra-web/pulls/124)).
+
 ### Fixed
 
 - Fixed broken custom emojis in bio when it is wrapped in `<div>` tag.